+++ packages/create-sitecore-jss/src/templates/nextjs/scripts/generate-component-builder/index.ts
@@ -1,10 +1,11 @@
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const plugins = require('scripts/temp/generate-component-builder-plugins');
-import { PackageDefinition } from '@sitecore-jss/sitecore-jss-dev-tools';
+import { PackageDefinition, ComponentFile } from '@sitecore-jss/sitecore-jss-dev-tools';
 
 export interface ComponentBuilderPluginConfig {
   watch?: boolean;
   packages?: PackageDefinition[];
+  components?: ComponentFile[];
 }
 
 export interface ComponentBuilderPlugin {