+++ packages/create-sitecore-jss/src/templates/nextjs/scripts/generate-component-builder/plugins/packages.ts
@@ -20,14 +20,6 @@ class PackagesPlugin implements ComponentBuilderPlugin {
      *    ]
      *  }
      */
-
-    /**
-     * Or you can register all components from a path using the below approach:
-     * import { PackageDefinition, getComponentList } from '@sitecore-jss/sitecore-jss-dev-tools';
-     * ...
-     * const componentsPath = 'src/extra';
-     * config.packages = getComponentList(componentsPath) as PackageDefinition[];
-     */
     config.packages = [];
 
     return config;