+++ packages/create-sitecore-jss/src/templates/nextjs/src/lib/next-config/plugins/cors-header.js
@@ -1,0 +1,22 @@
+﻿const config = require('../../../temp/config');
+
+/**
+ * @param {import('next').NextConfig} nextConfig
+ */
+const corsHeaderPlugin = (nextConfig = {}) => {
+  if (!config.sitecoreApiHost) {
+    return nextConfig;
+  }
+  return Object.assign({}, nextConfig, {
+    async headers() {
+      return [
+        {
+          source: '/_next/:path*',
+          headers: [{ key: 'Access-Control-Allow-Origin', value: config.sitecoreApiHost }],
+        },
+      ];
+    },
+  });
+};
+
+module.exports = corsHeaderPlugin;